// ui/src/lib/types/api.types.ts
export interface components {
    schemas: {
// ...
        CalculatorDto: {
            /** Format: uuid */
            id?: string;
            code?: string;
            name?: string;
            description?: string;
            active?: boolean;
            /** Format: date-time */
            createdAt?: string;
        };
        CreateCalculatorRequest: {
            code: string;
            name: string;
            description?: string;
            algorithmMetadata: components["schemas"]["FinancialProductDefinitionDto"];
        };
        CalculatorVersionDto: {
            /** Format: uuid */
            id?: string;
            /** Format: uuid */
            calculatorId?: string;
            version?: number;
            algorithmMetadata?: components["schemas"]["FinancialProductDefinitionDto"];
            /** Format: date-time */
            createdAt?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}